Move menu related functions to menu file
