+++ app/src/ui/lib/conflicts/unmerged-file.tsx
@@ -10,7 +10,7 @@ import {
 import { join } from 'path'
 import { Repository } from '../../../models/repository'
 import { Dispatcher } from '../../dispatcher'
-import { showContextualMenu } from '../../main-process-proxy'
+import { showContextualMenu } from '../../../lib/menu-item'
 import { Octicon } from '../../octicons'
 import * as OcticonSymbol from '../../octicons/octicons.generated'
 import { PathText } from '../path-text'
